refactor: centralize base64 decoding
